+++ app/Http/Controllers/Admin/NurseController.php
@@ -8,6 +8,9 @@ use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\User;
 use App\Http\Requests\UserFormRequest;
+use App\Record;
+use App\Care;
+use App\Plant;
 
 class NurseController extends Controller
 {
@@ -32,7 +35,7 @@ class NurseController extends Controller
             'description' => $request->description,
             'role' => 2
         ]);
-        return redirect()->route('nurse.index')->with(['alert' => 'Đã thêm thành công!']);
+        return redirect()->route('nurse.index')->with(['alert' => 'Đã thêm thành công!', 'type' => 'callout-success']);
     }
     public function show($id){
         $nurse = User::findOrFail($id);
@@ -49,12 +52,28 @@ class NurseController extends Controller
             'address' => $request->address,
             'password' => ($nurse->password == $request->password)? $nurse->password : bcrypt($request->password),
         ]);
-        return redirect()->route('nurse.index')->with(['alert' => 'Đã cập nhật thành công!']);
+        return redirect()->route('nurse.index')->with(['alert' => 'Đã cập nhật thành công!', 'type' => 'callout-success']);
     }
     public function destroy($id){
         $nurse = User::findOrFail($id);
+        $alert = [
+                    'alert' => 'Không thể xóa vì điều dưỡng viên này có liên quan tới một hoạt động khám hoặc điều trị!',
+                    'type' => 'callout-danger'
+                ];
+        $record = Record::select('*')->where('examiner', '=', $id)->first();
+        if(!is_null($record)){
+            return redirect()->route('nurse.index')->with($alert);
+        }
+        $plan = Plant::select('*')->where('user_id', '=', $id)->first();
+        if(!is_null($plan)){
+            return redirect()->route('nurse.index')->with($alert);
+        }
+        $care = Care::select('*')->where('staff_id', '=', $id)->first();
+        if(!is_null($care)){
+            return redirect()->route('nurse.index')->with($alert);
+        }
         $nurse->delete();
-        return redirect()->route('nurse.index')->with(['alert' => 'Đã xóa thành công!']);
+        return redirect()->route('nurse.index')->with(['alert' => 'Đã xóa thành công!', 'type' => 'callout-success']);
     }
     public function getSearch(Request $request){
         $query = $request->q;
